Map Seascape raster

Map Seascape raster

map_seascape_raster(
  r,
  palette = "Spectral",
  basemap = providers$Esri.OceanBasemap,
  opacity = 0.8
)

Arguments

r
raster layer

palette
color palette. Default: "Spectral".

basemap
basemap for leaflet. Default: leaflet::providers$Esri.OceanBasemap.

opacity
transparency. Default: 0.8.

var
variable. One of "CLASS" (default) or "P" for probability.

Value

Leaflet leaflet interactive map widget

Examples

r_monthly <- get_seascape_data(ctr_lon = -81.3, ctr_lat = 24.5, ctr_dd = 10, dataset = "global_monthly", var = "CLASS", date_beg = "2020-11-01", date_end = "2020-12-01")

#> Warning: The date_end 2020-12-01 > Seascapes end (2020-11-15) so decreasing to 2020-11-15.

#> info() output passed to x; setting base url to: https://cwcgom.aoml.noaa.gov/erddap/

map_seascape_raster(r_monthly)

#> Warning: Discarded ellps WGS 84 in Proj4 definition: +proj=merc +a=6378137 +b=6378137 +lat_ts=0 +lon_0=0 +x_0=0 +y_0=0 +k=1 +units=m +nadgrids=@null +wktext +no_defs +type=crs

#> Warning: Discarded datum World Geodetic System 1984 in Proj4 definition

#> Warning: Discarded ellps WGS 84 in Proj4 definition: +proj=merc +a=6378137 +b=6378137 +lat_ts=0 +lon_0=0 +x_0=0 +y_0=0 +k=1 +units=m +nadgrids=@null +wktext +no_defs +type=crs

#> Warning: Discarded datum World Geodetic System 1984 in Proj4 definition

#> Warning: Some values were outside the color scale and will be treated as NA

r_8day <- get_seascape_data(ctr_lon = -81.3, ctr_lat = 24.5, ctr_dd = 10, dataset = "global_8day", var = "CLASS", date_beg = "2020-11-01", date_end = "2020-12-01")

#> info() output passed to x; setting base url to: https://cwcgom.aoml.noaa.gov/erddap/

r_8day

#> class      : RasterStack 
#> dimensions : 401, 401, 160801, 5  (nrow, ncol, ncell, nlayers)
#> resolution : 0.05, 0.05  (x, y)
#> extent     : -91.35, -71.3, 14.5, 34.55  (xmin, xmax, ymin, ymax)
#> crs        : +proj=longlat +datum=WGS84 +no_defs 
#> names      : CLASS_2020.10.31, CLASS_2020.11.08, CLASS_2020.11.16, CLASS_2020.11.24, CLASS_2020.12.02 
#> min values :                3,                3,                3,                3,                3 
#> max values :               27,               28,               27,               27,               27 
#> 

r_monthly

#> class      : RasterLayer 
#> dimensions : 401, 401, 160801  (nrow, ncol, ncell)
#> resolution : 0.05, 0.05  (x, y)
#> extent     : -91.35, -71.3, 14.5, 34.55  (xmin, xmax, ymin, ymax)
#> crs        : +proj=longlat +datum=WGS84 +no_defs 
#> source     : memory
#> names      : CLASS 
#> values     : 3, 27  (min, max)
#> 

map_seascape_raster(raster::raster(r_8day, 1))

#> Warning: Discarded ellps WGS 84 in Proj4 definition: +proj=merc +a=6378137 +b=6378137 +lat_ts=0 +lon_0=0 +x_0=0 +y_0=0 +k=1 +units=m +nadgrids=@null +wktext +no_defs +type=crs

#> Warning: Discarded datum World Geodetic System 1984 in Proj4 definition

#> Warning: Discarded ellps WGS 84 in Proj4 definition: +proj=merc +a=6378137 +b=6378137 +lat_ts=0 +lon_0=0 +x_0=0 +y_0=0 +k=1 +units=m +nadgrids=@null +wktext +no_defs +type=crs

#> Warning: Discarded datum World Geodetic System 1984 in Proj4 definition

#> Warning: Some values were outside the color scale and will be treated as NA

map_seascape_raster(raster::raster(r_8day, 2))

#> Warning: Discarded ellps WGS 84 in Proj4 definition: +proj=merc +a=6378137 +b=6378137 +lat_ts=0 +lon_0=0 +x_0=0 +y_0=0 +k=1 +units=m +nadgrids=@null +wktext +no_defs +type=crs

#> Warning: Discarded datum World Geodetic System 1984 in Proj4 definition

#> Warning: Discarded ellps WGS 84 in Proj4 definition: +proj=merc +a=6378137 +b=6378137 +lat_ts=0 +lon_0=0 +x_0=0 +y_0=0 +k=1 +units=m +nadgrids=@null +wktext +no_defs +type=crs

#> Warning: Discarded datum World Geodetic System 1984 in Proj4 definition

#> Warning: Some values were outside the color scale and will be treated as NA