Plot Seascape Class as SVG

This function wraps around plot_ss_class to write the ggplot as a vector format file (*.svg) and return a path to the file that could be used for the web.

plot_ss_class_svg(var, val, g_svg, web_svg, redo = F, ht_ratio = 0.3, ...)

Arguments

var
variable to plot, for getting distribution of all Class averages from ss_gl_classes

val
average value to plot as a vertical line

g_svg
local file path to output ggplot as scalable vector graphic (*.svg)

web_svg
web prefix with which to return the path in HTML as an image

ht_ratio
ratio of height to width; default = 0.3

...
other parameters to pass onto plot_ss_class

Value