Plot Seascape Class Variable

Plot the provided value relative to all average values for the given variable across Seascape Classes.

plot_ss_class_var(
  var,
  val,
  n_bins = NULL,
  ply_color = "black",
  ply_alpha = 0.5,
  ln_color = "red",
  ln_size = 5,
  tbl_classes = ss_gl_classes
)

Arguments

var

variable to plot, for getting distribution of all Class averages from ss_gl_classes

val

average value to plot as a vertical line

n_bins

number of bins for generating histogram. If NULL (default), then perform a density histogram curve.

ply_color

color of polygon representing distribution of all average values for var; default = "black"

ply_alpha

color of polygon representing distribution of all average values for var; default = 0.5

ln_color

color of vertical line; default = "red"

ln_size

width of vertical line; default = 5

tbl_classes

table of values from which to extract var; default = ss_gl_classes