Get Seascape grids from polygon and date range

Given a polygon and date range, fetch Seascape data and return a raster layer for a single date or raster stack if multiple dates found for given date range.

get_ss_grds(
  ss_info,
  ply,
  ss_var = "CLASS",
  date_beg = "2020-01-01",
  date_end = max(get_ss_dates(ss_info)),
  dir_tif = NULL
)

Arguments

ss_info
SeaScape ERDDAP info object, as returned by get_ss_info)

ply
polygon as spatial feature sf, as returned by get_url_ply or bbox_to_ply

ss_var
SeaScape variable. One of "CLASS" (default) or "P" for probability.

date_beg
date begin to fetch, as character ("2003-01-15") or Date (Date("2003-01-15")).

date_end
date end to fetch, as character ("2020-11-15") or Date (Date("2020-11-15")). Defaults to most recent date available from ss_info.

Value

Raster raster layer if one date, stack if more

Examples

ply  <- get_url_ply("mbnms")
ss_i <- get_ss_info()
grds <- get_ss_grds(ss_i, ply, date_beg = "2020-01-01")
grds

#> class      : RasterBrick 
#> dimensions : 48, 41, 1968, 11  (nrow, ncol, ncell, nlayers)
#> resolution : 0.05, 0.05  (x, y)
#> extent     : -123.15, -121.1, 35.5, 37.9  (xmin, xmax, ymin, ymax)
#> crs        : +proj=longlat +datum=WGS84 +no_defs 
#> source     : memory
#> names      : CLASS_2020.01.15, CLASS_2020.02.15, CLASS_2020.03.15, CLASS_2020.04.15, CLASS_2020.05.15, CLASS_2020.06.15, CLASS_2020.07.15, CLASS_2020.08.15, CLASS_2020.09.15, CLASS_2020.10.15, CLASS_2020.11.15 
#> min values :                7,               14,                7,               12,                7,               12,                7,                7,               12,               14,               14 
#> max values :               21,               21,               21,               21,               21,               21,               21,               21,               21,               21,               21 
#>