Table of Seascape Class with plots per variable

Produce a table of variable average values for the Seascape Class with a distribution plot relative to all Classes. This function wraps around plot_ss_class_var.

tbl_ss_class(class, dir_svg, web_svg, tbl_classes = ss_gl_classes, ...)

Arguments

class
integer identifier for Seascape Class

dir_svg
local filesystem directory where to store SVG files

web_svg
web prefix for referencing the SVG files

tbl_classes
table of values from which to extract var; default = ss_gl_classes

...
other parameters to pass onto plot_ss_class_var

Value

Examples

dir_svg  <- here::here("inst/svg")
web_svg  <- "../svg"
tbl_ss_class(1, dir_svg, web_svg)

#> <table class=" lightable-material lightable-striped lightable-hover" style='font-family: "Source Sans Pro", helvetica, sans-serif; margin-left: auto; margin-right: auto;border-bottom: 0;'>
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> Variable </th>
#>    <th style="text-align:center;"> Class Avg </th>
#>    <th style="text-align:center;"> Relative to All Classes </th>
#>    <th style="text-align:center;"> All Min </th>
#>    <th style="text-align:center;"> All Max </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> SST (°C)<br>_sea surface temperature in degrees Celsius_ </td>
#>    <td style="text-align:center;"> 5.08 </td>
#>    <td style="text-align:center;"> <img src="../svg/ss_cl1_sst.svg"> </td>
#>    <td style="text-align:center;"> 0.15 </td>
#>    <td style="text-align:center;"> 28.25 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> SSS (psu)<br>_sea surface salinity in Practical Salt Units_ </td>
#>    <td style="text-align:center;"> 34.18 </td>
#>    <td style="text-align:center;"> <img src="../svg/ss_cl1_sss.svg"> </td>
#>    <td style="text-align:center;"> 27.74 </td>
#>    <td style="text-align:center;"> 38.14 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> ADT (m)<br>_absolute dynamic topography in meters_ </td>
#>    <td style="text-align:center;"> -0.37 </td>
#>    <td style="text-align:center;"> <img src="../svg/ss_cl1_adt.svg"> </td>
#>    <td style="text-align:center;"> -1.15 </td>
#>    <td style="text-align:center;"> 1.10 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> ICE (%)<br>_ice as percent cover_ </td>
#>    <td style="text-align:center;"> 0.00 </td>
#>    <td style="text-align:center;"> <img src="../svg/ss_cl1_ice.svg"> </td>
#>    <td style="text-align:center;"> 0.00 </td>
#>    <td style="text-align:center;"> 90.00 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> CDOM (m^-1^)<br>_colored dissolved organic matter per square meter_ </td>
#>    <td style="text-align:center;"> 0.01 </td>
#>    <td style="text-align:center;"> <img src="../svg/ss_cl1_cdom.svg"> </td>
#>    <td style="text-align:center;"> 0.00 </td>
#>    <td style="text-align:center;"> 0.07 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> CHLA (mg m^-3^)<br>_Chlorophyll _a_ per cubic meter_ </td>
#>    <td style="text-align:center;"> 0.21 </td>
#>    <td style="text-align:center;"> <img src="../svg/ss_cl1_chla.svg"> </td>
#>    <td style="text-align:center;"> 0.04 </td>
#>    <td style="text-align:center;"> 2.09 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> NFLH (W m^-2^ µm^-1^ sr^-1^)<br>_normalized fluorescence line height_ </td>
#>    <td style="text-align:center;"> 0.08 </td>
#>    <td style="text-align:center;"> <img src="../svg/ss_cl1_nflh.svg"> </td>
#>    <td style="text-align:center;"> 0.01 </td>
#>    <td style="text-align:center;"> 0.24 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> NFLH:CHL<br>_normalized fluorescence line height chlorophyll component_ </td>
#>    <td style="text-align:center;"> 0.37 </td>
#>    <td style="text-align:center;"> <img src="../svg/ss_cl1_nflh-chl.svg"> </td>
#>    <td style="text-align:center;"> 0.05 </td>
#>    <td style="text-align:center;"> 0.79 </td>
#>   </tr>
#> </tbody>
#> <tfoot><tr><td style="padding: 0; " colspan="100%">
#> <sup></sup> <ul>
#> <li>CLASS: 1</li>
#> <li>NAME: NORTH ATLANTIC SPRING, ACC TRANSITION</li>
#> <li>LATITUDE: SUBPOLAR</li>
#> <li>DOMINANT HEMISPHERE: SOUTH</li>
#> <li>DOMINANT SEASON: SPRING-AUTUMN</li>
#> </ul>
#> </td></tr></tfoot>
#> </table>